VB -4121 - Move from Prison API to Alerts API for prisoner alerts
Move the dto over to alerts package and begin capturing fields from
alerts api call
